Reduce to a single pass
You collect all of the nodes into a big array and then "filtered" that
array based on whether or not the class was present in the classList.

Instead of doing both, it's possible to just visit each node once and
add it to the results array when it passes the proposition of having the
className in its classList.